requirements.txt
streamlit==1.44.1
pandas==2.3.0
bcrypt==4.3.0
fpdf==1.7.2
pillow==11.2.1
qrcode==8.2
reportlab==4.4.0
svglib==1.5.1
PyPDF2==3.0.1
streamlit-option-menu==0.4.0
streamlit-autorefresh==1.0.1
plotly==6.0.1
streamlit-modal==0.1.2
pdf2image==1.17.0
google-cloud-storage==3.1.0
python-dotenv==1.1.0
psycopg2-binary==2.9.10
cloud-sql-python-connector==1.18.2
pg8000==1.31.2
SQLAlchemy==2.0.38
boto3==1.40.24
google-auth==2.40.3
openpyxl==3.1.5
num2words==0.5.14
